Update deproject.py
Added an option for a shell-by-shell conf() of the parameters.  It may be easiest to just use logging in sherpa to track the error bars, but I have it return the results of each shell's conf() in a data structure.  It's in reverse order, starting with the outer shell.  If you run:
p = dep.conf(), you can then get the mins and maxes for the outer shell using p[0].parmins and p[0].parmaxes.