Docs: update to reflect new and updated functionality, support RTD
Move to the Read-The-Docs theme and MathJaX for inline mathematics.

The Read-The-Docs build uses the standalone version of Sherpa provided as a
conda package for this build. This means that the documentation can not be
generated at build time, since there is no support for the required XSPEC
models in this version of CIAO. Note that the Sherpa version is not
versioned, but there is a restriction to python 3.5.

Reorganize the documentation so that there are multiple pages.

We have not yet integrarated the standalone and CIAO versions of the
Sherpa documentation (and may never do so). For now switch to the
standalone version for the "sherpa.astro.ui" layer commands (via the
reference links) since they are more up-to-date (but less detailed)
than the CIAO ahelp versions.