Relax the NumPy and hence AstroPy range limits
With CIAO 4.12 now available in conda I feel it's safe to remove
the NumPy / AstroPy restriction.

This still uses the standalone 4.11.1 version. It could be updated
to use CIAO 4.12, which would be good because we could include XSPEC
tests, but the runtime would be longer (as significantly more to be
downloaded). Note that Sherpa standalone 4.12.0 should be being
released soon.