Bump minimum Sherpa version to 4.14.0
I am seeing some odd errors on RTD which I can not see are related
to Sherpa changes, but let's bump to 4.14.0 as the minimum as I
do not have time to test with older versions.

As part of this some old code that referenced ChIPS has been
expunged from the world.